Move SchemaErrorsAccordion to its own folder
+ give it its own export
