+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/errors_accordion/index.tsx
@@ -20,9 +20,8 @@ import {
   EuiTableRowCell,
 } from '@elastic/eui';
 
-import { EuiButtonEmptyTo } from '../react_router_helpers';
-
-import { TruncatedContent } from '../truncate';
+import { EuiButtonEmptyTo } from '../../react_router_helpers';
+import { TruncatedContent } from '../../truncate';
 
 import './schema_errors_accordion.scss';
 
@@ -31,8 +30,8 @@ import {
   ERROR_TABLE_ERROR_HEADER,
   ERROR_TABLE_REVIEW_CONTROL,
   ERROR_TABLE_VIEW_LINK,
-} from './constants';
-import { FieldCoercionErrors } from './types';
+} from '../constants';
+import { FieldCoercionErrors } from '../types';
 
 interface ISchemaErrorsAccordionProps {
   fieldCoercionErrors: FieldCoercionErrors;
