+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/errors_accordion/schema_errors_accordion.scss
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-.schemaFieldError {
+ .schemaFieldError {
   border-top: 1px solid $euiColorLightShade;
 
   &:last-child {
